+++ src/spring-cloud/HISTORY.md
@@ -4,6 +4,7 @@ Release History
 ---
 * `az spring-cloud app deploy` has new preview argument "--build-env" to specify build module and jvm version and so on.
 * Raise error when `az spring-cloud app deploy` setting "--target-modules" and "--runtime-version for enterprise tier.
+* Fix the jvm option clearance in enterprise tier.
 
 3.0.0
 ---