+++ src/spring-cloud/HISTORY.md
@@ -1,5 +1,11 @@
 Release History
 ===============
+3.0.1
+---
+* `az spring-cloud app deploy` has new preview argument "--build-env" to specify build module and jvm version and so on.
+* Raise error when `az spring-cloud app deploy` setting "--target-modules" and "--runtime-version for enterprise tier.
+* Fix the jvm option clearance in enterprise tier.
+
 3.0.0
 ---
 * New preview argument `az spring-cloud create` has new argument "--sku=Enterprise" to support Azure Spring Cloud Enterprise creation.